Adjust folder structure. Seperate commons and sections
