+++ src/App.jsx
@@ -1,4 +1,4 @@
-import { Footer } from "./components/Footer/Footer";
+import { Footer } from "./components/commons/Footer/Footer";
 import { Home } from "./pages/Home";
 
 export const App = () => {
