--- src/components/Footer/Footer.jsx
@@ -1,5 +1,0 @@
-import "./Footer.scss";
-
-export const Footer = () => {
-  return <></>;
-};
