+++ src/components/commons/Footer/Footer.jsx
@@ -1,0 +1,13 @@
+import "./Footer.scss";
+
+export const Footer = () => {
+  return;
+
+  <>
+    <a href="mailto:hey@pernillasterner.se"> hey@pernillasterner.se</a>
+    <div className="social-links">
+      <a href="https://www.linkedin.com/in/pernilla-sterner/">Linkedin</a>
+      <a href="https://github.com/pernillasterner">GitHub</a>
+    </div>
+  </>;
+};
