+++ src/components/sections/Intro/Intro.scss
