+++ src/components/sections/Projects/Projects.scss
