+++ src/components/sections/Tech/Tech.scss
