+++ src/index.scss
@@ -106,6 +106,8 @@ h4 {
 p {
   // margin-bottom: 1.25rem;
   line-height: 1.5em;
+  font-size: 1.1rem;
+  font-weight: 400;
 }
 
 section {
