+++ src/pages/Home.jsx
@@ -1,6 +1,6 @@
-import { Intro } from "../components/Intro/Intro";
-import { Projects } from "../components/Projects/Projects";
-import { Tech } from "../components/Tech/Tech";
+import { Intro } from "../components/sections/Intro/Intro";
+import { Projects } from "../components/sections/Projects/Projects";
+import { Tech } from "../components/sections/Tech/Tech";
 import "./Home.scss";
 
 export const Home = () => {
